Move Studio Dask stats out of runner
